small fixes all around
