+++ wman-backend/Wman.Test/Builders/LogicBuilders/PhotoLogicBuilder.cs
@@ -1,10 +1,8 @@
 ﻿using CloudinaryDotNet.Actions;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Http.Internal;
 using MockQueryable.Moq;
 using Moq;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using Wman.Data.DB_Models;
@@ -71,19 +69,6 @@ namespace Wman.Test.Builders.LogicBuilders
             });
 
             return pictureList;
-        }
-
-        public static FormFile GetFormFile()
-        {
-            var content = "Hello World from a Fake File";
-            var fileName = "test.png";
-            var stream = new MemoryStream();
-            var writer = new StreamWriter(stream);
-            writer.Write(content);
-            writer.Flush();
-            stream.Position = 0;
-
-            return new FormFile(stream, 0, stream.Length, "id_from_form", fileName);
         }
 
         public static Mock<IPhotoService> GetPhotoService()
